4. maticar/maticar-backend/requirements/common.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.73.0
fastapi-mail==1.0.4
fastapi-pagination==0.9.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httptools==0.4.0
httpx==0.19.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
overrides==6.1.0
packaging==21.3
passlib==1.7.4
pbr==5.8.1
pdfkit==1.0.0
phonenumbers==8.12.45
Pillow==9.0.1
pip-review==1.1.1
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
redis==4.1.4
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
sqlalchemy-migrate==0.13.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
starlette==0.17.1
Tempita==0.5.2
typing-utils==0.1.0
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.7
websockets==10.2
wrapt==1.14.0
